rename SC_Ableton_Link->SC_LinkClock
