+++ lang/LangPrimSource/SC_LinkClock.hpp
